`rule_type lint`: This adds a `medev` command to lint rules
This is useful for us to start enforcing best practices when writing rules. Ideally, we should
start running this in CI wherever possible.

The initial implementation just adds regal [1] to help us lint rego rules.

[1] https://github.com/StyraInc/regal
