+++ go.mod
@@ -37,6 +37,7 @@ require (
 	github.com/spf13/viper v1.16.0
 	github.com/sqlc-dev/pqtype v0.3.0
 	github.com/stretchr/testify v1.8.4
+	github.com/styrainc/regal v0.9.0
 	github.com/xeipuuv/gojsonschema v1.2.0
 	go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc v0.45.0
 	go.opentelemetry.io/otel v1.19.0
